Nut: add support for 64-bits ints and for doubles.
The ULiL PS4 nut files are compiled with 64-bits ints, so this change is needed to parse them.
This isn't enough yet, because:
- In SQInstruction, I used SQInteger instead of int32_t, and switching to 64-bits SQIntegers breaks the SQInstruction parsing. This will be fixed in the next commit.
- The SQString length is a SQInteger, not a uint32_t. This code is shared with the ACT code, so I need to check the ULiL PS4 files before fixing that.

Also:
- print-act-nut switches from the "log and forget" error model to the exception error model. Nothing changes in the library, but the tool will no longer print thousands of lines.
I should completely remove the "log and forget" error model, always throw exceptions from the C++ API, and just catch them from the C API. And even in the C API, I think throwing exceptions should be safe as long as the caller is built with the same compiler and flags as the callee (which should be the case in tcrap), so I could make that catching a compile-time flag.
- Improved logging around the checkTag function.