SOuterType: use SQint_t as the internal type.
This one is hard, because it doesn't depend on the SQInteger size. Its backing type... is just whatever the compiler decided when the squirrel compiler is built - unlike SQInteger which uses the _SQ64 define. I guess we can assume the compiler will choose the target native type (32-bits for 32-bits targets and 64-bits for 64-bits targets), but I'm not sure the standard enforces that. And someone could still compile as 32-bits and define _SQ64 manually (even if I'm not sure the Squirrel team officially supports that).