Act: add support for Layer::weirdarray.
Reverse-engineered only by looking at 2 example files, but it manages to parse every PS4 file, so I guess it works...
I wish I had a proper name for that. But the only thing I know is that it's an array, and is isn't like other arrays, so that name will do for now.